x/battleship/types/message_set_field_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/verstakgit/battleship/testutil/sample"
)

func TestMsgSetField_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSetField
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSetField{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSetField{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
